requirements.txt
flask
flask_sqlalchemy
flask_oauthlib
openai
opencv-python-headless
numpy
python-dotenv